Refactored, decoupled functions, and added size considerations to the BULK call
