+++ src/playground/sample_data_ingestion/sample_data_ingestor/tests/test-files/valid-template-inputs.json
@@ -1,0 +1,42 @@
+{
+    "jinput": {
+        "year": "year",
+        "random number": "integer",
+        "email": "email"
+    },
+
+    "jinput2": {
+        "properties": {
+            "first name":    { "type" : ["first_name"]},
+            "last name":     { "type" : "last_name"},
+            "companies":{ "type" : ["array", "company", 5] }
+        }
+    },
+
+    "jinput3": {
+        "year": "year",
+        "addresses": ["array", "address", "integer", [1, 3]],
+        "director": "name"
+    },
+
+    "jinput4": {
+        "properties": {
+            "year":    { "type" : ["year"]},
+            "addresses":     { "type" : ["array", "address", "integer", [1, 3]] },
+            "director":{ "type" : "name" }
+        }
+    },
+
+    "jinput5": {
+        "properties": {
+            "year":    { "type" : ["year"]},
+            "integers":     { "type" : ["array", "integer", "integer", [1, 3], 5, 10] },
+            "director":{ "type" : "name" }
+        }
+    },
+
+    "jinput6": {
+        "date": "unix_time",
+        "average cpu usage": ["integer", 20, 30]
+    }
+}
