--- src/playground/sample_data_ingestion/utils/sample_data_ingestor/__init__.py
@@ -1,3 +1,0 @@
-# from sample_data_ingestor.sample_data_ingestor import ingest
-# from utils.sample_data_generator.sample_data_generator import choose_field
-# from utils.sample_data_generator.sample_data_generator import generate_data
